koniec

+++ mainCode/fromOutputToMainSite.py
@@ -9,7 +9,7 @@ with open(os.path.join(MAIN_DIR, "config.json"), 'r') as jsoner:
     shutil.copytree(os.path.join(OUTPUT_DIR, "imgs"),os.path.join(location, "imgs"))
     files = ["index.html", "main.js", "styles.css", "none.png"]
     for file in files:
-        if os.path.exists(os.path.join(location, file)):
+        if os.path.exists(os.path.join(location, file)) and os.path.exists(os.path.join(OUTPUT_DIR, file)):
             os.remove(os.path.join(location, file))
         if os.path.exists(os.path.join(OUTPUT_DIR, file)):
             shutil.copyfile(os.path.join(OUTPUT_DIR, file), os.path.join(location, file))
